Introduction of function AssertOnAndClearRTError()
This function checks if there is an RTE pending and if yes then ASSERTs.
If there was no pending RTE the function does nothing.
It should be used to ensure that no pending RTEs are present when entering
a critical section of code, like try / catch / endtry or any other
part that uses a local treatment of a RTE condition.