RemoveTracesFromGraph: Optimized RemoveFromGraph loop
Previously it was executed in a try/catch/endtry construct where the debugger
was disabled before and reenabled afterwards. This was slow and can be written
shorter.

The debugger setting change was removed and the error is now read back with
clear by GetRTError(1) on the same line. This prevents the debugger from
opening on error and also the err result can be checked to fall through
the while check.