Add a TP cycle ID
This allows to find out if testpulses in TPStorage or in StoredTestPulses
belong to the same TP run.

This is modelled after the RAC Id, see 6b92d92b (DA_Ephys/Labnotebook:
Introduce a "repeated acquisition cycle ID", 2017-07-12).

We do consider fast TP restarting to be part of the same cycle.