GetLastSettingChannel: Make returning INDEP_HEADSTAGE entries work again
In 83073ceb (GetLastSettingChannelInternal: Don't return a result for another valid channel, 2021-05-19) we fixed a bug where GetLastSettingChannelInternal returned NaN if another channel had a valid entry.

That fix broke support for headstage independent entries. This is now fixed.

The logic for handling indepdent entries was also move to
GetLastSettingChannelInternal as it can only be decided there properly.