PGC_SetAndActivateControl: Require a valid entry for Popup Menues
Until now we did only partially require that the index/string value for
Popup Menues is in the valid range.

This is done to prevent errors like 0ec7d31f (DA_Ephys: Use the correct GUI procedure
for Popup_Settings_DecMethod, 2021-09-29) to go unnoticed for more than
two years. This bug could now not happen anymore as now only strings which are
available in the popupmenu can be set.

In addition we also check the indizes.