Tests/TPDuringDAQOnlyTPWithLockedIndexing: Make the test work again without setting disabled controls
The test was added in 40956d16 (DA_Ephys: Make unlocked indexing with only
TP during DAQ channels working, 2021-01-22) to check that TP during DAQ
channels still allow to use DAQ with locked indexing even when this is the
only active one.

That approach tried to enable indexing only *after* DAQ was started. But
that approach was flawed as PGC_SetAndActivateControl just silently
ignored the request as the indexing controls are disabled during DAQ.

With the recent fix that PGC_SetAndActivateControl bugs out when trying to
set a disabled control, this no longer works.

The fix is a bit more involved as we now need to use the search box and
set the indexing end stimset to None. The indexing checkboxes are now also
set via the default mechanism using InitDAQSettingsFromString.