CONF_RestoreControl: Explicitly skip disabled controls
The recent change that PGC_SetAndActivateControl bugs out when used for a
disabled control breaks configuration as we are setting quite a few
controls those enabled/disabled state depends on other controls.

Now we have a couple of options:
1.) Pass mode == PGC_MODE_SKIP_ON_DISABLED to PGC_SetAndActivateControl to get the old
    behaviour
2.) Pass PGC_MODE_FORCE_ON_DISABLED to set the always set the control
3.) Enhance the experiment configuration logic to know when setting a
    disabled control does not make sense

The last option is the most work. And would also always need adaptation
when we introduce new business logic.

The second option makes it difficult to reason about the behaviour and
also has the potential to introduce inconsistent GUI states.

We therefore choose option 1 as that also matches the current behaviour.