DAP_SwitchSingleMultiMode: Clean it up
This function does multiple things at once. It handles the dependent
control disabling/enabling and also reenables the controls after DAQ when
required.

The former is now generalized and moved into AdaptDependentControls. The
latter can then be cleaned up a lot, as all remaining callers first enable
the CONTROLS_DISABLE_DURING_DAQ_TP and the let DAP_SwitchSingleMultiMode
adapt, and even set, them again.

But we only need to enable the controls again if we are in single device
mode.

And as the function now does something different we also rename it.