DAEphys: Add automatic TP and baseline adjustment
The new feature auto TP adjusts the baseline and the amplitude of the
testpulse to just make it work.

The amplitude setting in the DAEphys panel is now also per headstage,
the baseline percentage is still for all headstages.

The algorithm for both amplitude and baseline is independent but always
works on the same testpulses. The amplitude is only adapted once the
baseline is good. Once both are good for TP_AUTO_TP_CONSECUTIVE_PASSES
TPs, auto TP is automatically turned off for that headstage.

For the baseline adaptation we need two testpulse waveforms, this means we also auto-enable "Save TPs" for auto TP.
The waveforms are fetched via TP_GetStoredTPs and belong therefore the
same testpulse cycle.

The "TP buffer size" labnotebook entry is now not saved anymore as part
of data acquisition but as part of the testpulse settings.

New labnotebook entries:
- TP Auto On/Off
- TP Auto max current
- TP Auto voltage
- TP Auto voltage range
- TP buffer size (INDEP_HEADSTAGE)
- Minimum TP resistance for tolerance (INDEP_HEADSTAGE)
- Send TP settings to all headstages (INDEP_HEADSTAGE)
- TP Auto percentage (INDEP_HEADSTAGE)
- TP Auto interval (INDEP_HEADSTAGE)
- TP Auto QC
- TP Cycle ID