SF: Fix in SF_ACTION_ARRAY handling
SweepFormula handled double arrays like [[1]] wrong and returned always
single arrays in the form of [1]. The issue was that the recursive calling
for array evaluation did not track if the parsing of the inner buffer
implicitly already created an array or not.
This is not tracked and properly handled.

Also added some in-code comments for SF_ACTION_ARRAY.

Created separate functions for specific JSON handling.