IsConstant: Allow to not ignore NaNs and fix edge cases
The issue triggering this [1] claimed that IsConstant does not work with
NaN in the input.

But actually the current behaviour was just to ignore NaNs in the input,
which also all callers relied on.

We now have the option to not ignore NaNs in the input.

We also fixes edge cases for empty input waves (return NaN) and the case
were all input is NaN and we ignore NaN. The latter now gives NaN as this is
equivalent to an empty input wave.

[1]: https://github.com/AllenInstitute/MIES/issues/1124

Close #1124.