PA_RetrievePulseInfosFromEpochs: Fix pulse retrieval
The parsing code incorrectly assumed that the length entry always comes
before the pulse active entry, but that is not guaruanteed.

For a correct correlation of the total pulse length and the active
start/stop times, we also need to parse the pulse number from the epoch
info. Doing that requires also to keep track of the epoch numbers as the pulse
number start from zero on every epoch.

Bug present since ba209bbd (PA plot: Gather more pulse infos, 2021-02-02).