SC_SpikeControl: Enhance reaction to failed too many spikes QC
The current approach is to reuse the existing daScaleModifier and
daScaleOperator with a slight twist as in

SetDAScaleModOp(panelTitle, i, 2 * daScaleModifier, daScaleOperator, invert = 1)

which should allow to get the number of spikes down as

SetDAScaleModOp(panelTitle, i, daScaleModifier, daScaleOperator)

is used for too few spikes.

But real world experiments have shown this does not work.

So we now use the new mandatory analysis parmeters
DaScaleTooManySpikesOperator and DaScaleTooManySpikesOperator for the
former case.