PSQ analysis functions: Enforce that we have the correct sampling frequency
The new optional "SamplingFrequency" analysis parameter (defaulting to
50kHz) is used for checking in POST_SWEEP_EVENT that the data was
acquired with the correct sampling interval.

As PSQ_Chirp was missing the neighbouring "SamplingMultiplier" this was
added as well.

Not having the correct sampling interval is a QC failure. This is
documented in the new PSQ_FMT_LBN_SAMPLING_PASS labnotebook entry.