PSQ_EvaluateBaselineProperties: Rework test override integration
Since the first version of the patch seq analysis functions in efdfbadf
(Analysis function for Patch Seq subthreshold stimsets, 2017-09-20) we
used the approach to only override the chunkPassed result for the tests.

While this works great in general, this has the drawback that now the
presence of the targetV labnotebook entries (voltage and QC state)
depends on the actual rmsShortPassed/rmsLongPassed QC results measured
with the hardware.

This is unfortunate for the tests, as this means that sometimes the
target voltage entries are present in the labnotebook and sometimes not.

By using chunkPassed from test override as value for
rmsShortPassed/rmsLongPassed/targetVPassed we avoid that situtation.

This also allows us to get rid of the test override logic for the return
values which always looked odd to have at all.

This has no observable behaviour change for real users.