FindLevelWrapper: Preserve wave layout of input data with layers
The FindLevelWrapper accepts zero or one layer as wave dimension of data.

As we need to redimension data to be 1D wave, we have to set it back
after the level operations.

But we missed the fact that the input could have a layer.

Fix that, and also add a test to check it.