Fix for PSQChirp function to not skip baselineQC check on spikecheck fail
Previously the chirp analysis function aborted early when a the spike
detection concluded a fail. Tthe baselineQC analysis was not executed.
This prevented proper chunk entries in the epoch information.

This was wrong, as on the data that was actually acquired the
baselineQC check is required to generate the proper information in the
LBN and Epochs.
With this change the baselineQC analysis is run as well, also creating
proper chunk entries in the epochs.