SweepFormula Help File: Allow showing it again on experiment load with open databrowser
We have the SweepFormula Help file shown as a notebook without saving
it's content in the recreation macro. This is done so that newer
versions of the help file are loaded from disc.

We currently only load the SweepFormula help file on
Databrowser/Sweepbrowser creation.

This results in the bug that opening an existing experiment with a DB/SB
never has a filled SweepFormula help notebook.

Fix that by always checking if the help file is empty on tab change.