WB_SaveStimSet: Create the stimulus set before calling AFH_CheckAnalysisParameter
The new signature of the analysis function check functions has the
stimulus set name included. Users therefore assume that the set already
exists and WB_CreateAndGetStimSet returns a wave.

But that is not the case.

Fix it. In case the check function returns an error, we just delete the
stimulus set again.