PSQ_Chirp: Add bounds evaluation mode
Feedback from real world users has shown that the current approach to
look at the minimum and maximum of the sine chirp makes the test very
difficult to pass.

We therefore add the possibility to only look at the Maximum or only the
Minimum. This can be controlled with the new BoundsEvaluationMode
analysis parameter.