TS_GetNewestFromThreadQueue/TS_GetNewestFromThreadQueueMult/TS_ThreadGroupGetVariable: Avoid debugger popup
In some cases we do have a race condition between the fifo thread on
data acquisition stopping itself in TFH_FifoLoop with TFH_STOP_ACQ mode
and the background functions querying the thread queue.

On developer machines with debugging and debug on error turned on, this
can result in a debugger popup.

Avoid that by using the err = ... trick available since Igor Pro 7.