TP_GetValuesFromTPStorage: Handle TPStorage with only one entry properly
lastValidEntry holds the index of the last row with data. So in case we
request only one entry we now return $"". But that is wrong as we have
to compare numReqEntries with lastValidEntry + 1 as that is the number
of valid entries. And as that is even more confusing we introduce
numValidEntries.

In addition we need to return with valid results from outside the loop
as we only run the loop once in this case.