DAEphys: Avoid creating the ITC1600 Device 0 path for yoking if not needed
Yoking is a specialized feature not every user is using. But we always
do create the full device path including the ListOfFollowerITC1600s
global string.

But we actually only need to do that if we have a ITC1600 Device 0 at
all.

All other code paths of GetFollowerList only call it when yoking is
done.