Databrowser/SweepBrowser: Enable/Disable Passed/Failed checkboxes with Dashboard checkbox
The current behaviour is to enable the Passed/Failed checkboxes of the
dashboard if we have sweeps and disable them otherwise.

But this is not how we usually do things. In addition it also hinders
the user changing the checkboxes while no sweeps have been acquired.

With this patch, we now toggle the enabled state of these checkboxes
with the Dashboard enable checkbox.