SweepBrowser/SettingsHistory: Return the experiments from the current instance
We used to return all experiments available in the analysis browser in
each the Settings History popup menu.
But this is confusing as users expect only the experiments which they
have loaded into the SweepBrowser to be present.

Fix that.

This fix also has the nice side effect that
LBV_SelectExperimentAndDevice now preselects the experiment if all
sweeps from the SweepBrowser are from the same experiment.