DC_Configure: Adapt miscellaneous waves clearing
We currently clear in DC_Configure and related functions various
waves usind during DAQ/TP.

In a future commit we want to write into the settings wave from all
analysis function events and this requires us to clear those wave for
DAQ even earlier as DC_Configure.

Clearing the waves is now done in DC_ClearWaves. For TP we call it as
usual from DC_Configure. But for DAQ we call it from the acquisition
state handling code as that also catches cases where we restart DAQ
without calling DAP_CheckSettings.