PSQ_NumPassesInSet: Always return an integer
In the dashboard we use this function to gather the passes in the set
during data acquisition.

Therefore it can happen that the Sweep QC state of a sweep is still NaN.
In this case PSQ_NumPassesInSet would always return NaN.

This is not what we want so we now default to returning zero when
querying the Sweep QC state for missing entries.

For the PSQ analysis function this change does not change anything as
all analysis functions call this function with Sweep QC labnotebook
entries present.