HW_ITC_CloseDevice: Only close devices which can be selected
Ever since we got stricter in 3be3c285 (DAP_UnlockDevice: Abort on
hardware errors related to device unlocking, 2021-11-24) we now get
error messages to the history when unlocking a DAEphys panel were the
ITC device is already closed.

We only can close the device if it is active, therefore we don't need to
continue unless it is active.