DC_MakeHelperWaves: Workaround stale wave locks
For the analysis function calls in AFM_CallAnalysisFunctions we lock the
passed waves so that the analysis function can not change the wave
accidentally.

But if these locks are kept for whatever reasons, users can currently
not start data acquisition again.

Let's warn the user with a BUG message in this case and clear the lock.