AD_FillWaves: Handle not executing analysis functions gracefully
Until now we always search for analysis function results if we have found
that a given headstage has an entry for it.

But the global "Skip analysis function calls" checkbox allows to not
execute analysis functions at all. In this case we still have the labnotebook
entries with the analysis function entry present.

Modify the outputted result in this case, so that the user knows that it
was just not run but still set.