AD_GetResultMessage: Handle manually stopped sets better
When stopping a sweep in a set we currently don't issue
POST_SWEEP_EVENT/POST_SET_EVENT events. Which means that some
labnotebook entries like set QC are not written. This means we can not
treat a missing set QC entry as "data acquisition is ongoing".

By looking at the "DAQ stop reason" we can figure out if the set was
prematurely stop.