AdditionalExperimentCleanup: Handle reopening the debug panel correctly
We want to close all panels/graphs/tables except the background watcher
and the debug panel.

But as the debug panel has a listbox which references waves in
root:MIES, the listbox is invalid if we temporarily move these waves
around.

We now therefore close the debug panel, move the list waves, and reopen
it.

Close #1233.