Make the TP power spectrum work again with odd number of rows
The FFT operation used in SCOPE_UpdatePowerSpectrum requires that the
OscilloscopeData wave has an even number of rows.

The easiest way to ensure that is to make it have even number of rows in
the first place in DC_GetStopCollectionPoint and also make
OscilloscopeData hold an even number of entries.

The bug is present since forever but is only now more visible since
the default baseline changed in da36d5a8 (DAEphys: Use 35% as default
value for the baseline percentage, 2021-11-10) as since then the TP
length is not even in more cases.

But it turns out that for combinations like 2 HS, 1000ms, 35% baseline
the FFT operation takes so long as to result in an ITC crash as the FIFO
overflows.

The reason is that for the above combination the number of points is
833334 which has the following prime numbers 2, 3, 138889 and the used
FFT algorithm is *very very* slow for large primes.

In order to fix that we pad with zeros until the next power of 2. This
is also much faster even as the number of points transformed is much
larger.

Close #1232.