ED_FindIndizesAndRedimension: Optimize LBN_SetDimensionLabels calls
At least since 5f77b077 (Move dimension label setting logic into its own
function, 2014-12-21) we set the dimension labels on every call to
ED_createTextNotes/ED_createWaveNotes.

This is actually a complete waste of time, as we only add labnotebook
entries. So it should be enough to initially set the dimension labels and
then only set the new ones.

We do that now.