Tests/TestEpochsGeneric: Fix invalid unit conversion
Sampling interval is stored in ms. This also requires to make the
endTimeEpochs check differently as it fails now on EP_EpochTest6.

By just making the calculated endTimeDAC time one samplingInterval
larger we get the same logic as was originally envisioned in e5c450b5
(Tests for Epoch information gathering on various stimsets, 2019-06-25).