Tests/TestEpochsGeneric: Revert lastPoint fix
In 8c6e5da6 (EP_WriteEpochInfoIntoSweepSettings: Handle unacquired data,
2021-07-13) we patched this code in order to handle unacquired epochs.

But the change is actually wrong as we want to be sure that all epochs
lie inside the DAC wave and don't care about the unacquired epoch
specially.

The reason why that went unnoticed is that the last commit hid that bug
by rounding endTimeEpochs to 1 and therefore the assertion
CHECK_LE_VAR(endTimeEpochs, endTimeDAC) always passed.