EP_AdaptEpochInfo: Properly size the unacquired epoch
Since 8c6e5da6 (EP_WriteEpochInfoIntoSweepSettings: Handle unacquired
data, 2021-07-13) we store the unacquired data as additional epoch.

The algorithm to determine the range of that epoch is to use the
acquiredTime (last time point with acquired data) as starting point. And
the maximum of all encountered epochs as ending point.

This works as long as all of the looked at epochs lie inside the sweep
initially. This is the case for stock epochs. But not necessarily for user epochs.

By using the last point of the sweep as ending point we avoid issues
with user epochs which are too long.

This bug was found while playing around with the PSQ_Ramp analysis
function as that adds an user epoch which is as long as
stopCollectionPoint since f063c50c (PSQ_Ramp: Add user epochs for
special AD/DA ranges, 2021-09-20).