SCOPE_ITC_AdjustFIFOPos: Return the correct length for fifoPos == NaN
The new PSQ_RAMP test PS_RA2a has a passing pre pulse baseline, failing
post pulse baselines and a spike. This results in DAQ stopping and
starting when the spike was found, and this makes the offset from
GetDataOffset() non-zero, and data acquisition runs till the very end.

Now due to how TFH_FifoLoop works the fifo position (relative to the
offset) it pushes into the queue, is always clamped between 0 and
stopCollectionPoint (borders included). Only once data acquisition is
finished, TS_GetNewestFromThreadQueue returns NaN as the FIFO thread can
not be reached anymore, see for example DQM_FIFOMonitor.

This means SCOPE_ITC_UpdateOscilloscope is called twice at the end of
data acquisition. Once with fifoPos being stopCollectionPoint and after
that with fifoPos being NaN.

And this now triggers a bug in SCOPE_ITC_AdjustFIFOPos. On the
second call, with fifoPos being NaN, it returns stopCollectionPoint - offset. As the
global fifo position is already stopCollectionPoint, the check

if(fifoPosGlobal == fifoPos)
        return NaN
endif

in SCOPE_ITC_UpdateOscilloscope does not trigger and we try to do

scaledDataWave[fifoPosGlobal, fifoPos - 1] = ...

with `fifoPosGlobal == stopCollectionPoint` and `fifoPos == stopCollectionPoint - offset`.

Now scaledDataWave has exactly stopCollectionPoint rows, therefore we
immediately get a wave indexing error.

This bug is present since c01f4660 (Introduce scaledDataWave,
2019-09-11). But it only surfaces since c6468854 (Introduction of function
AssertOnAndClearRTError(), 2021-09-16) as we don't ignore lingering RTEs
anymore.

The author of this commit believes that no data corruption is present
in PSQ_RAMP data due to this bug as the indexing into scaledDataWave
bugs outs immediately.