Tests/BUG: Allow tests to expect BUG messages
Since 4a364db9 (Make calling the BUG() function during hardware tests a CI
failure, 2021-04-01) we treat bugs during testing as asssertions and also
as CI failure.

This works great in general, but in certain scenarios it can be beneficial
to expect bug messages in tests and still pass them.
This is now possible.