Labnotebook entries: Remove a.u. as units
Arbitrary unit, short a.u., is a concept to denote an unit to something
which has only meaning relative to a reference value, see also [1].

But in most of the current use cases we don't have these kind of values,
instead it was mostly used to denote "no unit".

For that we now just use the empty string.

[1]: https://en.wikipedia.org/wiki/Arbitrary_unit