GetLBNumericalValues: Add description entry handling
As we now have for all stock numerical LBN entries additional information
available, we also want to store that in the key wave.

The extensive checking for consistent entry metadata in
ED_FindIndizesAndRedimension will help in finding inconsistencies when
adding new LBN entries.

The numerical keys wave has now 6 rows; new are the Description,
HeadstageContigency and the ClampMode.

Although we don't yet have a description wave for the textual entries, we
already resize it, so that both key waves have a consistent number of
rows.