CreateTiledChannelGraph: Fix assertion with dDAQ slider
When acquiring oodDAQ data and when trying to view the data with dDAQ
checkbox enabled and AUTOMATED_TESTING being defined, we get an assertion from
EnforceIndependentSetting introduced in d0dfc2e8 (Labnotebook: Enforce
that the INDEP getters are correctly used, 2021-04-22).

We can avoid that by querying the INDEP_HEADSTAGE entry of "Stim set
length" only when dDAQ is enabled, as it is in the headstage dependent
entries for all other cases.