Databrowser/SweepBrowser: Add option to show only selected sweeps in logbook graph
Since forever the default x-axis range for the logbook graph is
autoscaled. This means that all entries are shown, which can be quite
unsavy when you have a lot of sweeps acquired.

We now allow with the new checkbox in the settings history panel to
restrict the x-axis to the selected sweeps.