GetKeyWaveParameterAndUnit: Rework it and enhance it
The current implementation uses pass-by-reference, and also returns the
parameter name, which always conincides with the passed key name. It also
does not use FindValue's RMD flag, probably because it was implemented
back in 4caea03f (Add a new graph to show data from the settingsHistory vs
time, 2014-11-18) which targetted IP6 where this flag was not available.

This is now all fixed. And also tests were added.