SF_OperationTP: Read out the clamp amplitude from the tags
For the inserted TP this is the same as the stored value in the LBN, see
DC_FillDAQDataWaveForDAQ and there the globalTPInsert part. It is
important to note that DAScale is not applied to the inserted TP.

But for user defined TP's, which are just parts of a normal stimset, the
DAScale value is applied on top of the amplitude of the stimset epoch.

So by using the "Amplitude" entry of the TP_PX epoch we have a common
base for both types, stock and user ones.