WB_SaveStimSet: Avoid deleting existing stimsets on errors
Since ed8b9434 (WB_SaveStimSet: Create the stimulus set before calling
AFH_CheckAnalysisParameter, 2021-11-18) we delete the just created
stimset in case the checks fail.

This approach results in an existing stimset being deleted in case it
could not be saved.

We don't want that. So instead we now create a temporary stimset name,
check that and only if the checks pass save it under the desired name.