Tests: Enhance stimset loading speedup heuristics
Since 439b21f6 (Hardware Tests: Make them faster locally, 2021-04-01)
we don't reload the stimsets from disc if they are already loaded.

This allows for faster test execution on developer machines.

But that approach failed if the stimset on disc changed.

This is now fixed by looking at the modification date of the stimset
file.

As this speedup heuristics is not used on CI machines, it is also not
critical.