Tests: Check amplifier presence before locking
We currently get a crash when the MultiClamp Commander is not open and
the hardware tests are run with an ITC1600.

While the reason for the crash is difficult to find out, we can at least
avoid it by checking that the amplifier is present before device
locking.

Introducing the routine EnsureMCCIsOpen also avoids code duplication.