Tests: Make RestoreDAEphysPanel work with real amplifier
Since the first version of that test in e39315e0 (Tests for Load/Save of
GUIState and Experiment state through DAEphys panel, 2019-11-13) the
configuration JSON referneced the demo amplifier (serial 0).

This means the test fails with a real amplifier.

Let's fix that by replacing the used serial with the available one.