POST_SWEEP_EVENT/POST_SET_EVENT: Add support for writing user epochs
Up to now it was not possible to add new user epochs in the
POST_SWEEP_EVENT/POST_SET_EVENT.

The reason was that at this point we already wrote the epoch information
from the epochs wave to the labnotebook. So any subsequent addition of
new epochs to the epochs wave was ignored.

This is now fixed. The approach is that, after storing the epoch
information in the labnotebook in SWS_SaveAcquiredData, we fill the
epochs wave from the labnotebook again and after the POST_SET_EVENT
write it into the labnotebook again.

In this way analysis functions can now read and write user epochs from
the POST_SWEEP_EVENT/POST_SET_EVENT.

The EP_TestUserEpochs was also adapted to check for the newly supported
events, and also acquires data with RA so that the POST_SET_EVENT
triggers.