tools/unit-testing: Check JUNIT xml files
We have a report [1] that failing tests with ITC1600 hardware results in
invalid JUNIT files due to invalid UTF8-characters.

A preliminary analysis suggests that the problem is some kind of memory
corruption.

In order to debug that issue, we now check all generated JUNIT XML files
during CI are XML-compliant.

[1]: https://github.com/byte-physics/igor-unit-testing-framework/issues/236