DC_FillDAQDataWaveForTP: Fix reusing incompatible NI channel waves
Since 878eeaf9 (DC_PlaceDataInDAQDataWave: Make the unchanged TP case
faster, 2020-12-21) we reuse existing DAQDataWaves when running TP.
Although we tried to ensure that all settings were the same, this failed
for NI hardware as the did not check the separate NI channel waves.

By reading and writing the TP_PROPERTIES_HASH recursively, we now ensure
that the DAQDataWave and the separate channel waves (NI hardware only) are
matching.