Tests for analysis functions: Add generic range checks for labnotebook entries
We already check a lot of labnotebook entries for each test. This works as
long as the stored value has a defined value. But some values are just in
some known range.

By checking the range, derived from the unit, we can ensure that we have
the correct magnitude of the entry.