LBV_CheckIfXAxisIsTime: Tweak defaults without traces
In 253baf7a (LBV_CheckIfXAxisIsTime: Tweak default value depending on existing sweeps, 2021-11-16)
we added some logic to select the time axis without acquired sweeps.

We can do better as we actually always want time axis for TPStorage, and
we also use default time axis for the results waves.

And therefore the call to GetPlainSweepList is not needed anymore.