tools: Skip deployment when it is outdated
Since forever our deployment scripts for the documentation and the release
assets just deployed what they were given (publish-docs-on-github.sh) or
returned an error if HEAD did not resolve to a branch
(upload-github-release-asset.sh).

HEAD not resolving to a branch can happen if we merge two PRs in a row.
When in this case two plans for deployment are started, the first one will
see HEAD being something older than latest main (or release branch) [1].

This was seen a couple of times over the years but never properly
understood.

By silently ignoring the deployment we do the right thing and also
don't fail the job anymore.

[1]: http://bamboo.corp.alleninstitute.org/browse/MIES-BM-305/