SF Tests: Adapt data operation tests using select
The last test for requesting more sweeps than there are existing was adapted
due to select returning only existing sweeps. For this very special case this
a functionality change of data that uses now select comapred to data with
channels, sweeps specified.

However returning empty sweeps before for non-existing sweeps was inconvenient
and could be considered as wrong.