PSQ_CR_DetermineBoundsAction: Use correct conversion in production
We need to multiply by 1000 for converting from V to mV. This is in the
production-code only branch of the code, so could not be found during
testing.

The bug is present since c371535c (Add PSQ_Chirp, 2020-06-02). It is
unclear how a bug like that could pass manual testing. The commit message
of feaf34d8 (PSQ_Chirp: Add bounds evaluation mode, 2021-11-19) hints that
getting a passing sweep in production is very difficult.