SF: Optimization for speed of select operation.
Previously we determined all active channels from all sweeps before we
resolved the actual active channels of each individial sweep.
This is inefficient as we can as well resolve first the actual channels of all
sweeps. Then for displayed sweeps we can filter that list based on the
trace data.

Also in SF_CreateResultsWaveWithCode we can use the channels data determined
by the select() execution. Select() returns the data of the displayed sweeps,
so the returned channels are also only from the displayed sweeps.