SF: Test fix TestSweepFormulaCodeResults test with data from store operation
Previously TestSweepFormulaCodeResults saved the sweeps from the select
operation and the channels from the channels operation. This was wrong,
because "select()" works on the displayed sweeps and channels on the
given channels. If the displayed sweeps did not contain any of the given
channels, channels were saved that did not exist in the sweeps that were saved.

Now a single select() operation retrieves both. However, since active channels
are saved separately from the sweeps there might be active channels that are
not active for a particular sweep (and other channels thar are no active for
another sweep).

This is summarized in a current issue 1335