GetUniqueEntries: added dontDuplicate optional argument and Tests
GetUniqueEntries was very slow if the input wave has only a single element
because it created a new free wave. In a lot of application cases it suffices
to continue using the input wave instead. For theses cases the duplicate can
be prevented.

The new optional argument dontDuplicate allows to hint the function that it
is ok to return the input wave if it has only a single element preventing the
mem alloc overhead of duplicate.