Wavebuilder: Use newer PRNG for new stimsets
We want to use the new Xoshiro PRNG for new stimsets, while using the
old PRNGs for existing ones.

This is achieved by introducing
SegmentParameters::noiseGenMode/noiseGenModePTMixedFreq which hold the
old PRNG values when the wavebuilder parameter wave is upgraded but
NOISE_GEN_XOSHIRO when it is newly created.

GetReproducibleRandom now also defaults to NOISE_GEN_XOSHIRO, the call
sites which care about reproducibility are converted to us the
noiseGenMode.