PSQ_EvaluateBaselineProperties: Add new mode
We now have a new mode PSQ_AVERAGE_TEST which calculates the average
voltage in each baseline chunk and stores it in the labnotebook.

There is no QC step involved.

An interesting detail here is that we do call
PSQ_EvaluateBaselineProperties for the new analysis function
PSQ_TRUE_REST_VM during midsweep and therefore we have to pass
the epochsWave to EP_GetEpochs as the labnotebook entries don't yet
exist.