EP_GetEpochs: Allow querying epochs during mid sweep event
Since forever we only allowed querying the epoch information from the
labnotebook. This approach has the limitation that we can't query epoch
information during mid sweep event. But we already have the correct
format of the already added epoch info in the wave from GetEpochsWave.

We therefore allow to pass that wave in.