DB: Add tooltip help and syntax coloring support for sweepformula notebook
The following features were added:

- on activate/deactivate of the notebook simple syntax coloring is applied that
  colors operation keywords.
- When marking an operation with shift + rightclick the user can jump to the
  sf help notebook to the operation description
- When marking an operation then a tooltip with the help is shown where the
  operation help text is extracted from the sf help notebook.

Developer notes: As notebooks can not have their own tooltips the tooltip is
registered to the BrowserSettingsPanel. It triggers when the mouse is over the
SF_InfoTab that contains the notebook. Thus, there always has to be a control
that covers the area behind the notebook.