Adapted Sweepformula help notebook for required convention
The help logic for the sweepformula notebook requires to have a
specific convention for the logic to work.

The operation title must be "operation - <operatioName>" and the name of the
special character notebook action "to top" at the end of the operation section
must be to_top_<operationName>. Name refers here to the internal name of the
action, not the shown string/title.