PSQ_SealEvaluation: Make BaselineChunkLength a lower limit for WaveBuilder epoch length
Since it's introduction in 35176831 (PSQ_EvaluateBaselineProperties: Add
support for new method, 2022-02-22) and 1a64ca17 (PSQ_SealEvaluation:
Add it, 2022-02-22) one can select epochs for baseline QC by creating a
baseline chunk selection user epoch (U_BLSX).

We currently require that the WaveBuilder epoch exactly matches the
given BaselineChunkLength.

This is too limiting. Instead we now require that the WaveBuilder epoch
is at least as long as the BaselineChunkLength.

The code PSQ_EvaluateBaselineProperties already only looks at the user epoch
and not at the length of the WaveBuilder epoch.