SF: Fix cursors operation to use the executor for argument parsing
In the cursors operation the arguments were parsed manually through a direct
JSON getter. Now the arguments are correctly parsed through the FormulaExecutor

It was also incorrectly working for zero arguments returning a zero sized wave.
As zero arguments for cursors were not explicitly specified, the code was
adapted to default to cursors(A,B) for that case.