PSQ_PipetteInBath: Implement sweep formula execution more robust
We used to need fiddling with OVS so that our Sweep Formula execution
works, as we could not reference non-displayed sweeps back then we
implemented it in 6e32699f (PSQ_PipetteInBath: Add it, 2021-12-08).

This is now fixed.

The testcase PS_PB3 is also now more correct as the SweepFormula is
executed although the baseline did not pass and the sweep finished early.