LBV_UpdateLBGraphLegend: Tweak legend
At the very beginning of the labnotebook in MIES we treated the last
layer as the "all" entry for all headstages. This was very quickly
abandoned in favour of the concept of having headstage independent
entries stored there instead.

We therefore now rename the legend in the logboook viewer from all to indep.