SF: Change + operation to use permanent wref waves in working df
- adapted to plotter to handle wref waves correctly as input
  currently only the first element is considered.
- add define SWEEPFORMULA_DEBUG that enables that all data waves are stored as
  permanent waves in the working datafolder. The names are prefixed to
  indicate where the wave was created.

Within the Formula Executor the waveref waves are stored as single element
text waves. The element stores the full path to a permanent wave reference
wave. The element is prefixed to differentiate between an actual single
element text wave from e.g. a user input.
User input are returned by the FormulaExecutor as regular data waves.
If such input is encountered it is automatically converted to a wave ref
input.