PSQ_CreateOverrideResults: Add missing break statement
This resulted in the override wave of PSQ_PipetteInBath having the same
size as the one for PSQ_SealEvaluation.

As the size of the former is always larger than the latter one, this only
wasted memory, but did not had any visible bug.

Broken since 1a64ca17 (PSQ_SealEvaluation: Add it, 2022-02-22).