SF: Fix for select only returning data for sweeps with existing singleSweepDF
When using the select operation it checks if the required DF for a sweep is
present. Though, the single sweep DF for a sweep is only created if it is
shown in the Databrowser. Thus, when using mode "all" and a requested sweep
was not shown before its single sweep DF was not created and select skipped
these sweeps.