PSQ_Chirp: Limit repurposed time to 6s
The stimset for PSQ_Chirp can be very long. It therefore can currently
happen that by repurosing the left over sweep time plus 10s, the ITI is
actually very large.

As this wastes a lot of time, and is also not necessary, let's limit the
ITI to arounds 6s even if the sweep is longer.