FitResistance: Avoid warning message if it does not make sense
Since 6a43fbcd (PatchSeqSubThreshold: Calculate the resistance of a passed
sweep, 2017-09-27) the analysis function PatchSeqSubThreshold, now called
PSQ_DAScale in sub-threshold mode, uses the plotting mechanism for the
resistance.

This plotting mechanism, introduced in 92087e3c (Add analysis function for
stepping through a DA Scale range, 2017-09-08), warns the user if one of
the labnotebook entries deltaI or deltaV could not be found.

This only makes sense if these entries are present for each sweep. But
for PSQ_DAScale these are only present if the sweep passed.

So let's not warn if the sweep failed in this case.